@@ -61,7 +61,7 @@ export default {
       }
     }, 500);
   },
-  destroyed() {
+  unmounted() {
     clearInterval(this.interval);
   },
 };
