@@ -1,32 +1,30 @@
-/*
- Copyright 2017-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// import './init'
+// import { createApp } from 'vue/dist/vue.esm-bundler';
+// import App from './App.vue'
+// import 'vuetify/styles'
+// import { createVuetify } from 'vuetify'
+// import * as components from 'vuetify/components'
+// import * as directives from 'vuetify/directives'
+// const { aliases, mdi } = require('vuetify/iconsets/mdi');
+// import VuexStore from './store'
+// import { createStore } from 'vuex';
+// import router from './router'
 
- Licensed under the Amazon Software License (the "License"). You may not use this file
- except in compliance with the License. A copy of the License is located at
+// const vuetify = createVuetify({
+//   components,
+//   directives,
+//   icons: {
+//     defaultSet: 'mdi',
+//     aliases,
+//     sets: {
+//       mdi,
+//     },
+//   },
+// })
 
- http://aws.amazon.com/asl/
+// const app = createApp(App)
 
- or in the "license" file accompanying this file. This file is distributed on an "AS IS"
- BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, express or implied. See the
- License for the specific language governing permissions and limitations under the License.
- */
-
-/**
- * Entry point for sample application.
- * See lex-web-ui.js for the component entry point.
- */
-
-import Vue from 'vue';
-import router from '@/router';
-import LexApp from '@/LexApp';
-
-/* eslint-disable no-new */
-new Vue({
-  router,
-  render: (h) => h(LexApp),
-}).$mount('#lex-app');
-
-Vue.config.errorHandler = (err, vm, info) => {
-  // eslint-disable-next-line no-console
-  console.error('unhandled error in lex-app: ', err, vm, info);
-};
+// app.use(vuetify)
+// app.use(router)
+// app.use(createStore(VuexStore))
+// app.mount('#lex-app')
